// lib/enum.ts
export const enum Query {
  eq = 'eq',
  neq = 'neq',
  gt = 'gt',
  gte = 'gte',
  lt = 'lt',
  lte = 'lte',
  in = 'in',
  nin = 'nin',
  all = 'all',
  size = 'size',
  exists = 'exists',
  type = 'type',
  regex = 'regex',
  and = 'and',
  or = 'or',
  nor = 'nor',
  not = 'not',
  limit = 'limit',
  offset = 'offset',
  select = 'select',
  sort = 'sort',
}